pyproject.toml
[project]
name = "grid-sort-benchmark"
version = "0.1.0"
description = "2D grid-sorting algorithm benchmark (LAS, GradSort, RasterFairy, SOM, KS)"
requires-python = ">=3.10"

dependencies = [
    "matplotlib",
    "numpy>=1.24",
    "pillow>=10.0",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "scikit-image>=0.21",
    "torch>=2.1",
    "torchvision>=0.16",
    "minisom>=2.3",
    "rasterfairy @ git+https://github.com/Quasimondo/RasterFairy.git@035ac797f27413ef4356e070aedd52797a0c43ce",
    "isomatch @ git+https://github.com/Sarmagan/IsoMatch-Python.git",
]

[tool.uv]
# Package not on PyPI — install manually or place the folder in the project root:
#   KernelizedSorting_master : place the folder in the project root